Add mopp import/export and parse command

+++ TagTool/Geometry/BlenderPhmoMoppUtil.cs
@@ -87,7 +87,7 @@ namespace TagTool.Geometry
                 DataBuildType = -101
             };
 
-            resource.DataCapacityAndFlags = resource.DataSize;
+            resource.DataCapacityAndFlags = (uint)resource.DataSize;
 
             return resource;
         }
